# docs/swagger.yaml
definitions:
  dto.CategoryCreateDTO:
    properties:
      budget:
        type: number
      name:
        type: string
      priority:
        type: integer
    type: object
  dto.CategoryDTO:
    properties:
      budget:
        type: number
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      note_id:
        type: string
      priority:
        type: integer
      updated_at:
        type: string
    type: object
  dto.CategorySetBudgetDTO:
    properties:
      budget:
        type: number
    type: object
  dto.NoteCreateDTO:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.NoteDTO:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    required:
    - name
    - user_id
    type: object
  dto.TransactionDTO:
    properties:
      amount:
        type: number
      category_id:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      label:
        type: string
      note_id:
        type: integer
      timestamp:
        type: string
      updated_at:
        type: string
    type: object
  dto.UserLoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.UserRegisterDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      password_confirmation:
        type: string
    required:
    - email
    - name
    - password
    - password_confirmation
    type: object
  models.Category:
    properties:
      budget:
        type: number
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      note:
        $ref: '#/definitions/models.Note'
      note_id:
        type: string
      priority:
        type: integer
      updated_at:
        type: string
    type: object
  models.Note:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.Transaction:
    properties:
      amount:
        type: number
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      label:
        type: string
      note:
        $ref: '#/definitions/models.Note'
      note_id:
        type: integer
      timestamp:
        type: string
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: User
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginDTO'
      responses:
        "200":
          description: OK
      summary: Login a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegisterDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
      summary: Register a user
      tags:
      - auth
  /notes:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.NoteDTO'
            type: array
      security:
      - BearerAuth: []
      summary: List notes
      tags:
      - notes
    post:
      consumes:
      - application/json
      parameters:
      - description: Note
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/dto.NoteCreateDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Note'
      security:
      - BearerAuth: []
      summary: Add a note
      tags:
      - notes
  /notes/:note_id/transactions:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
      security:
      - BearerAuth: []
      summary: List transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      parameters:
      - description: transaction
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/dto.TransactionDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Transaction'
      security:
      - BearerAuth: []
      summary: Add a transaction
      tags:
      - transactions
  /notes/{note_id}:
    get:
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NoteDTO'
      security:
      - BearerAuth: []
      summary: Get a note
      tags:
      - notes
  /notes/{note_id}/categories:
    get:
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CategoryDTO'
            type: array
      security:
      - BearerAuth: []
      summary: List categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: string
      - description: Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.CategoryCreateDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CategoryDTO'
      security:
      - BearerAuth: []
      summary: Add a category
      tags:
      - categories
  /notes/{note_id}/categories/{category_id}/budget:
    post:
      consumes:
      - application/json
      parameters:
      - description: Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.CategorySetBudgetDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryDTO'
      security:
      - BearerAuth: []
      summary: Set budget for a category
      tags:
      - categories
  /users:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"